fixed compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@8020 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ filters/include/pcl/filters/impl/bilateral.hpp
@@ -101,7 +101,7 @@ pcl::BilateralFilter<PointT>::applyFilter (PointCloud &output)
     tree_->radiusSearch ((*indices_)[i], sigma_s_ * 2, k_indices, k_distances);
 
     // Overwrite the intensity value with the computed average
-    output.points[(*indices_)[i]].intensity = computePointWeight ((*indices_)[i], k_indices, k_distances);
+    output.points[(*indices_)[i]].intensity = static_cast<float> (computePointWeight ((*indices_)[i], k_indices, k_distances));
   }
 }
